% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{active}
\alias{active}
\title{Currently active configurations}
\usage{
active()
}
\value{
A character vector with the name(s) of currently active configurations
}
\description{
Determine the currently active configurations.
}
\details{
The names of currently active configurations are read from the
\code{R_CONFIG_ACTIVE} environment variable (configuration names
can be space or comma delimited). If the variable is not defined then
"default" is returned.
}
\seealso{
\link{is_active}
}

