% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Verify config contains all reference entries}
\usage{
check(config, reference)
}
\arguments{
\item{config}{A configuration object (list)}

\item{reference}{A reference configuration object (list)}
}
\value{
TRUE if all reference entries are in the config list, FALSE otherwise.
}
\description{
Verify config contains all reference entries
}
