% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{config}
\alias{config}
\title{Manage Environment Specific Configuration Values.}
\description{
Manage configuration values across multiple environments (e.g. development,
test, production). Read values using a function that determines the current
environment and returns the appropriate value.
}
\details{
The main function is \code{\link[=get]{get()}}.

For additional details see \url{https://rstudio.github.io/config/}.
}
