% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_config}
\alias{extract_config}
\title{Extract element from config list (any level)}
\usage{
extract_config(conf, key)
}
\arguments{
\item{conf}{A configuration list}

\item{key}{The key to extract}
}
\value{
A list or value from the configuration object with the name of \code{key}. NULL if key is not found.
}
\description{
Extract a value (can be a value or list) from a configuration list named by \code{key}.
}
\details{
With complex configurations, extracting a value from the configuration becomes a
challenge. There isn't a great general-purpose way that I could find for extracting a
list/value at an arbitrary depth of configuration. This function performs that function.

Do note that a list is not guaranteed to have unique names at different levels. This function
will return the object found at the shallowest level (from the initial object). If multiple
such matches are found, this will be noted but ignored.

Note that the function returns NULL if the key is not found
}
\examples{

extract_config(list(l11=list(
     l21=list(l31=letters, l32=rev(letters)),
     l22=toupper(letters)),
  l12=mtcars$mpg)
)
}
