% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filters}
\alias{filters}
\alias{add_filter}
\alias{remove_filter}
\title{Specify filters for configuration values}
\usage{
add_filter(filter)

remove_filter(handle)
}
\arguments{
\item{filter}{Filter function}

\item{handle}{Handle to previously installed filter function.}
}
\description{
Provide a filter that is called from the \code{\link[=get]{get()}} method with the resolved
configuration. The filter can modify the configuration and return an
alternate configuration.
}
\details{
The \code{add_filter()} function returns a handle which can be
subsequently passed to \code{remove_filter()} to remove the filter.
}
