% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.d.R
\name{get.d}
\alias{get.d}
\title{Get config from multiple individual files}
\usage{
get.d(dir, pattern = "*.conf", ...)
}
\arguments{
\item{dir}{The directory to use for config files.}

\item{pattern}{(default *.conf) File pattern for config files.}

\item{...}{Any other parameters to pass to \code{\link[=get]{get()}}.}
}
\value{
A configuration list (see \code{\link[=get]{get()}}).
}
\description{
This uses the linux approach of a \code{config.d} type directory with individual configs to load
and merge.
}
\examples{

\dontrun{
conf <- get.d(dir = "config.d")
}
}
