% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{is_active}
\alias{is_active}
\title{Test active configuration}
\usage{
is_active(config)
}
\arguments{
\item{config}{Configuration name(s)}
}
\value{
Logical vector indicating whether the specified configurations are active
}
\description{
Check whether a configuration is currently active
}
\seealso{
\link{active}
}

