% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge two configurations. Always use as \code{config::merge()}.}
\usage{
merge(base_config, merge_config)
}
\arguments{
\item{base_config}{Configuration to merge values into}

\item{merge_config}{Configuration to merge values from}
}
\value{
Configuration which includes the values from
\code{merge_config} merged into \code{base_config}.
}
\description{
Merge one configuration into another recursively.
}
\section{Warning - Do not attach the package using library(config)}{


We strongly recommend you use \code{config::get()} rather than attaching the
package using \code{library(config)}.

In fact, we strongly recommend you never use \code{library(config)}.

The underlying reason is that the \code{get()} and \code{merge()} functions in
\code{{config}} will mask these functions with the same names in base R.
}

\seealso{
\code{\link[=get]{get()}}
}
