% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save}
\alias{save}
\title{Save configuration values to file.}
\usage{
save(
  file = "config.yml",
  config_list = list(),
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  overwrite = FALSE
)
}
\arguments{
\item{file}{Configuration file to write to (defaults to
\code{"config.yml"}).}

\item{config_list}{List of configuration options to be saved to config file.}

\item{config}{Name of configuration profile to save. Defaults to
the value of the \code{R_CONFIG_ACTIVE} environment variable
("default" if the variable does not exist).}

\item{overwrite}{Should an existing config file be overwritten (default is FALSE)}
}
\value{
None (invisible NULL) as returned from \code{\link[base:cat]{base::cat()}}.
}
\description{
Save the currently active configuration from a list.
}
\details{
For additional details see \url{https://rstudio.github.io/config/}.
}
